commented out tests for deepl translation and dbdiff tests in common app. Fixed errors in common tests.py and included unittest execution of common tests in github actions workflow
