+++ webcentral/src/templates/pages/grid_listing.html
@@ -1,6 +1,6 @@
 {% extends 'base.html' %}
 {% load i18n %}
-{% block title %}{{ title }}{% endblock %}
+{% block title %}{{ heading }}{% endblock %}
 {% block content %}
   {% load static %}
