+++ webcentral/src/templates/partials/explanation_show_more.html
@@ -8,7 +8,7 @@
            alt="Circle icon"
            class="description-heading__icon" />
     {% endif %}
-    <p class="mb-0 description-heading__paragraph">{{ title }}</p>
+    <p class="mb-0 description-heading__paragraph">{{ heading }}</p>
   </div>
   <div class="description-content">
     <div class="mainContent col-lg-12">
